Rename BackgroundTranslator -> TranslatorUI.
