--- getting-started/background-processing/BackgroundTranslator/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using BackgroundTranslator
-@using BackgroundTranslator.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
