+++ getting-started/background-processing/TranslatorUI/Program.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace BackgroundTranslator
+namespace TranslatorUI
 {
     public class Program
     {
