+++ getting-started/background-processing/TranslatorUI/Startup.cs
@@ -11,7 +11,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace BackgroundTranslator
+namespace TranslatorUI
 {
     public class Startup
     {
