+++ getting-started/background-processing/TranslatorUI/TranslatorUI.csproj
