+++ getting-started/background-processing/TranslatorUI/Views/Home/Index.cshtml
