+++ getting-started/background-processing/TranslatorUI/Views/Home/Privacy.cshtml
