+++ getting-started/background-processing/TranslatorUI/Views/Shared/_ValidationScriptsPartial.cshtml
