+++ getting-started/background-processing/TranslatorUI/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using TranslatorUI
+@using TranslatorUI.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
