+++ getting-started/background-processing/TranslatorUI/Views/_ViewStart.cshtml
