+++ getting-started/background-processing/TranslatorUI/appsettings.Development.json
