+++ getting-started/background-processing/TranslatorUI/appsettings.json
