+++ getting-started/background-processing/TranslatorUI/wwwroot/js/site.js
